LPD-58844 Move mail service impl out into modules
