+++ modules/apps/mail/mail-messaging-impl/build.gradle
@@ -1,9 +1,11 @@
 dependencies {
+	compileOnly group: "com.liferay.jakarta.portlet", name: "com.liferay.jakarta.portlet-api", version: "4.0.0"
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "default"
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
 	compileOnly group: "jakarta.activation", name: "jakarta.activation-api", version: "2.1.3"
 	compileOnly group: "jakarta.mail", name: "jakarta.mail-api", version: "2.1.3"
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.4.0"
 	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
+	compileOnly project(":apps:portal:portal-aop-api")
 	compileOnly project(":core:petra:petra-string")
 }
