+++ modules/apps/mail/mail-messaging-impl/src/main/java/com/liferay/mail/messaging/internal/MailServiceImpl.java
@@ -3,14 +3,16 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.mail.service.impl;
+package com.liferay.mail.messaging.internal;
 
 import com.liferay.mail.kernel.auth.token.provider.MailAuthTokenProvider;
 import com.liferay.mail.kernel.auth.token.provider.MailAuthTokenProviderRegistryUtil;
 import com.liferay.mail.kernel.model.Account;
 import com.liferay.mail.kernel.model.MailMessage;
 import com.liferay.mail.kernel.service.MailService;
+import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
+import com.liferay.portal.aop.AopService;
 import com.liferay.portal.kernel.change.tracking.CTAware;
 import com.liferay.portal.kernel.cluster.Clusterable;
 import com.liferay.portal.kernel.jndi.JNDIUtil;
@@ -27,7 +29,6 @@ import com.liferay.portal.kernel.util.PrefsPropsUtil;
 import com.liferay.portal.kernel.util.PropertiesUtil;
 import com.liferay.portal.kernel.util.PropsKeys;
 import com.liferay.portal.kernel.util.PropsUtil;
-import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.Validator;
 
 import jakarta.mail.Authenticator;
@@ -44,11 +45,15 @@ import java.util.function.Function;
 import javax.naming.Context;
 import javax.naming.InitialContext;
 
+import org.osgi.service.component.annotations.Component;
+
 /**
  * @author Brian Wing Shun Chan
  */
+@Component(service = AopService.class)
 @CTAware
-public class MailServiceImpl implements IdentifiableOSGiService, MailService {
+public class MailServiceImpl
+	implements AopService, IdentifiableOSGiService, MailService {
 
 	@Clusterable
 	@Override
@@ -64,6 +69,13 @@ public class MailServiceImpl implements IdentifiableOSGiService, MailService {
 		}
 
 		_sessions.remove(companyId);
+	}
+
+	@Override
+	public Class<?>[] getAopInterfaces() {
+		return new Class<?>[] {
+			MailService.class, IdentifiableOSGiService.class
+		};
 	}
 
 	@Override
