--- portal-impl/src/META-INF/mail-spring.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0"?>
-
-<beans
-	default-destroy-method="destroy"
-	default-init-method="afterPropertiesSet"
-	xmlns="http://www.springframework.org/schema/beans"
-	xmlns:util="http://www.springframework.org/schema/util"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util.xsd"
->
-	<bean class="com.liferay.mail.service.impl.MailServiceImpl" id="com.liferay.mail.kernel.service.MailService" />
-</beans>
