+++ portal-impl/src/portal.properties
@@ -854,7 +854,6 @@
         /META-INF/repository-spring.xml,\
         \
         /META-INF/counter-spring.xml,\
-        /META-INF/mail-spring.xml,\
         /META-INF/portal-spring.xml,\
         /META-INF/portlet-container-spring.xml,\
         \
