+++ portal-kernel/src/com/liferay/mail/kernel/service/packageinfo
@@ -1,1 +1,1 @@
-version 6.0.0
+version 7.0.0
